move inventory performance profiles to compressed air assessment, update styling
