+++ src/app/compressed-air-assessment/compressed-air-assessment.module.ts
@@ -132,7 +132,7 @@ import { SharedCompressorPipesModule } from './shared-compressor-pipes/shared-co
 import { AssessmentCo2SavingsModule } from '../shared/assessment-co2-savings/assessment-co2-savings.module';
 import { CompressorSummaryTableModule } from './compressor-summary-table/compressor-summary-table.module';
 import { CompressorSummaryComponent } from './system-profile/compressor-summary/compressor-summary.component';
-import { InventoryPerformanceProfileModule } from '../shared/inventory-performance-profile/inventory-performance-profile.module';
+import { InventoryPerformanceProfileModule } from './inventory-performance-profile/inventory-performance-profile.module';
 import { WelcomeScreenComponent } from './welcome-screen/welcome-screen.component';
 import { Co2HelpTextModule } from '../shared/co2-help-text/co2-help-text.module';
