+++ src/app/compressed-air-assessment/compressed-air-report/compressed-air-report.module.ts
@@ -17,7 +17,7 @@ import { SharedCompressorPipesModule } from '../shared-compressor-pipes/shared-c
 import { PrintOptionsMenuModule } from '../../shared/print-options-menu/print-options-menu.module';
 import { CompressorSummaryTableModule } from '../compressor-summary-table/compressor-summary-table.module';
 import { SystemProfileGraphsModule } from '../system-profile-graphs/system-profile-graphs.module';
-import { InventoryPerformanceProfileModule } from '../../shared/inventory-performance-profile/inventory-performance-profile.module';
+import { InventoryPerformanceProfileModule } from '../inventory-performance-profile/inventory-performance-profile.module';
 import { PerformanceProfilesComponent } from './performance-profiles/performance-profiles.component';
 
 @NgModule({
