+++ src/app/compressed-air-assessment/inventory-performance-profile/inventory-performance-profile.component.css
@@ -1,0 +1,5 @@
+.group-label{
+    background-color: white;
+    padding-left: .5rem;
+    padding-top: .5rem;
+}
